Rename extension files
