+++ src/Microsoft.Health.Fhir.Anonymizer.Shared.Core/Extensions/TypedElementNavExtensions.cs
